requirements.txt
fastapi
uvicorn
pydantic
Pillow
sqlalchemy
sqlalchemy-utils
psycopg2-binary
pytest
python-dotenv